It compiles, but doesn't work properly.
Structure initialization can't work with this: if it was initialized as
an int then used later as a struct, none of the struct's fields are
there, so writting to them is not doable.

There might be a way, using `remove` and `initialize` to refresh the
register at every use. This would kill the side-channel attack used to
pass parameters, and so special care would have to be taken for them.